refactor: 알림 로직 형식 공통화
