+++ src/main/java/com/workhub/cs/service/csPost/UpdateCsPostService.java
@@ -8,7 +8,6 @@ import com.workhub.cs.entity.CsPostFile;
 import com.workhub.cs.entity.CsPostStatus;
 import com.workhub.cs.entity.CsQna;
 import com.workhub.cs.service.CsPostAccessValidator;
-import com.workhub.cs.service.CsPostNotificationService;
 import com.workhub.cs.service.csQna.CsQnaService;
 import com.workhub.global.entity.ActionType;
 import com.workhub.global.error.ErrorCode;
@@ -56,7 +55,7 @@ public class UpdateCsPostService {
         Set<Long> commenters = csQnaService.findAllByCsPostId(csPostId).stream()
                 .map(CsQna::getUserId)
                 .collect(Collectors.toSet());
-        csPostNotificationService.notifyCsPostUpdated(projectId, csPostId, csPost.getTitle(), commenters);
+        csPostNotificationService.notifyUpdated(projectId, csPostId, csPost.getTitle(), commenters);
 
         return CsPostResponse.from(csPost, visibleFiles);
     }
