+++ src/main/java/com/workhub/projectNotification/entity/NotificationType.java
@@ -11,11 +11,9 @@ public enum NotificationType {
     PROJECT_INFO_UPDATED,
     POST_CREATED,
     POST_UPDATED,
-    POST_DELETED,
     POST_COMMENT_CREATED,
     CS_QNA_CREATED,
     CS_QNA_ANSWERED,
     CS_POST_CREATED,
-    CS_POST_UPDATED,
-    CS_QNA_COMMENT_CREATED
+    CS_POST_UPDATED
 }
