+++ src/test/java/com/workhub/cs/service/csPost/CreateCsPostServiceTest.java
@@ -4,7 +4,6 @@ import com.workhub.cs.dto.csPost.CsPostFileRequest;
 import com.workhub.cs.dto.csPost.CsPostRequest;
 import com.workhub.cs.dto.csPost.CsPostResponse;
 import com.workhub.cs.entity.CsPost;
-import com.workhub.cs.service.CsPostNotificationService;
 import com.workhub.global.error.ErrorCode;
 import com.workhub.global.error.exception.BusinessException;
 import com.workhub.project.entity.Project;
